copy to clipboard added to onlick chart event
