+++ lib/index.js
@@ -16,6 +16,7 @@ import { extendDeep } from './utils/extend-deep';
 import exportImage from './utils/export-svg';
 import renderDownloadButton from './utils/render-download-btn';
 import renderExecutionMetadata from './utils/render-execution-metadata';
+import copyToClipBoard from './utils/copy-to-clipboard';
 
 import palettes from './palettes/index';
 
@@ -57,6 +58,9 @@ export const Dataviz = function (options = {}) {
     ui: {
       executionMetadata: true,
     },
+    utils: {
+      clickToCopyToClipboard: true, 
+    },
 
     indexBy: 'timeframe.start',
     labels: [],
@@ -93,9 +97,6 @@ export const Dataviz = function (options = {}) {
 
     axis: {},
     color: {},
-    data: {
-      order: null
-    },
     size: {
       // control it with CSS of .c3-chart
     },
@@ -120,11 +121,13 @@ export const Dataviz = function (options = {}) {
       // duration: 0
     },
     data: {
+      order: null,
       selection: {
        enabled: true,
        draggable: true,
        multiple: true
-     }
+     },
+     onclick: (d, element) => copyToClipBoard(d.value),
    },
    grid: {
      y: {
@@ -252,7 +255,22 @@ export const Dataviz = function (options = {}) {
     this.render(this.config.results);
   }
   this.config.firstVisibilityState = document.visibilityState;
+
+  if (!this.config.utils.clickToCopyToClipboard) {
+    this.config.data.onclick = () => {};
+  }
+
+  if (options.data && options.data.onclick) {
+    this.config.data.onclick = (d, element) => {
+      const callback = options.data.onclick.bind(null, d, element);
+      callback();
+
+      if (this.config.utils.clickToCopyToClipboard) copyToClipBoard(d.value);
+    };
+  }
+
 };
+  
 
 Dataviz.libraries = { default: {} };
 if (typeof window !== 'undefined') {
