+++ lib/utils/copy-to-clipboard.js
@@ -1,0 +1,8 @@
+export default function copyToClipBoard(value) {
+  const placeholder = document.createElement('input');
+  placeholder.value = value;
+  document.body.appendChild(placeholder);
+  placeholder.select();
+  document.execCommand('copy');
+  document.body.removeChild(placeholder);
+}
