+++ lib/utils/render-execution-metadata.js
@@ -1,3 +1,5 @@
+import copyToClipBoard from './copy-to-clipboard';
+
 function removeClassWithTimeOut(element, className = 'copied', ms = 1000) {
   setTimeout(() => {
     element.classList.remove(className);
@@ -23,12 +25,7 @@ function handleCopyToClipboard(event) {
     }
   }
 
-  const placeholder = document.createElement('input');
-  placeholder.value = value;
-  document.body.appendChild(placeholder);
-  placeholder.select();
-  document.execCommand('copy');
-  document.body.removeChild(placeholder);
+  copyToClipBoard(value);
 }
 
 export default function renderExecutionMetadata({
